게시판 CRUD 인증/인가 구현 (#21)
* feat : 게시글 작성 기능 인증 구현

* feat : 게시글 수정 기능 인증/인가 구현

* feat : 게시글 삭제 기능 인증/인가 구현

* feat : 게시글 작성자 조회 구현

* feat : 로그인 한 사용자 표시

+++ src/main/java/com/study/proudcat/domain/post/service/PostService.java
@@ -9,6 +9,8 @@ import com.study.proudcat.domain.post.dto.response.PostDetails;
 import com.study.proudcat.domain.post.entity.Post;
 import com.study.proudcat.domain.post.repository.PostRepository;
 import com.study.proudcat.domain.storage.repository.StorageRepository;
+import com.study.proudcat.domain.user.entity.User;
+import com.study.proudcat.domain.user.repository.UserRepository;
 import com.study.proudcat.infra.exception.ErrorCode;
 import com.study.proudcat.infra.exception.RestApiException;
 import com.study.proudcat.infra.utils.FileUtils;
@@ -20,10 +22,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -35,17 +34,15 @@ public class PostService {
     private final PostRepository postRepository;
     private final FileDataRepository fileDataRepository;
     private final StorageRepository storageRepository;
-
-
-    @Transactional
-    public void writePostTest(WritePostRequest request) {
-        postRepository.save(request.toEntity());
-    }
+    private final UserRepository userRepository;
 
     @Transactional
-    public void writePost(WritePostRequest request, MultipartFile image) throws IOException {
+    public void writePost(WritePostRequest request, MultipartFile image, String email) throws IOException {
         log.info("PostService writePost run..");
         log.info("upload file : {}", image.getOriginalFilename());
+
+        User writer = userRepository.findByEmail(email)
+                .orElseThrow(() -> new RestApiException(ErrorCode.NO_TARGET));
 
         FileData fileData = FileUtils.parseFileInfo(image);
         if (fileData == null) {
@@ -53,7 +50,7 @@ public class PostService {
         }
         fileDataRepository.save(fileData);
 
-        Post post = request.toEntity();
+        Post post = request.toEntity(writer);
         post.setFileId(fileData.getId());
 
         postRepository.save(post);
@@ -96,7 +93,6 @@ public class PostService {
     public List<FindPostResponse> getPostsSearchList(String title, Pageable pageable) {
         log.info("PostService getPostsSearchList run..");
         Page<Post> postPage = postRepository.findAllPostsPage(title, pageable);
-        List<FindPostResponse> responses = new ArrayList<>();
         return postPage.getContent()
                 .stream()
                 .map(FindPostResponse::from)
@@ -128,7 +124,7 @@ public class PostService {
     }
 
     @Transactional(readOnly = true)
-    public PostDetails getPostWithCommentsById(Long postId) throws IOException {
+    public PostDetails getPostWithCommentsById(Long postId) {
         log.info("PostService getPostWithCommentsById run..");
         Post post = getPostEntity(postId);
         return PostDetails.from(post);
@@ -141,16 +137,24 @@ public class PostService {
     }
 
     @Transactional
-    public void modifyPost(Long postId, ModifyPostRequest request) {
+    public void modifyPost(Long postId, ModifyPostRequest request, String email) {
         log.info("PostService modifyPost run..");
         Post post = getPostEntity(postId);
+
+        if (!post.isSameWriter(email)) {
+            throw new RestApiException(ErrorCode.NOT_PROPER_USER);
+        }
         post.modify(request.getTitle(), request.getDescribe());
     }
 
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, String email) {
         log.info("PostService deletePost run..");
         Post post = getPostEntity(postId);
+
+        if (!post.isSameWriter(email)) {
+            throw new RestApiException(ErrorCode.NOT_PROPER_USER);
+        }
         postRepository.delete(post);
     }
 
@@ -164,15 +168,5 @@ public class PostService {
     private Post getPostEntity(Long postId) {
         return postRepository.findById(postId)
                 .orElseThrow(() -> new RestApiException(ErrorCode.NO_TARGET));
-    }
-
-    private byte[] getByteFile(String filePath) throws IOException {
-        byte[] byteFile = null;
-        try {
-            byteFile = Files.readAllBytes(new File(filePath).toPath());
-        } catch (IOException e) {
-            System.out.println("File to byte[] 변환 에러");
-        }
-        return byteFile;
     }
 }
